pyproject.toml
[tool.poetry]
name = "WatchUI"
version = "2.0.0"
description = "RobotFramework library package for automated visual testing."
authors = ["Your Name <you@example.com>"]
license = "MIT"
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
]

[tool.poetry.dependencies]
python = "^3.9"
opencv-python = "*"
pandas = "*"
Pillow = "*"
numpy = "*"
scikit-image = "*"
scikit-learn = "*"
imutils = "*"
robotframework = "*"
selenium = "*"
robotframework-seleniumLibrary = "*"
robotframework-browser = "*"
PyMuPDF = "*"
pytesseract = "*"


[tool.poetry.dev-dependencies]
black = "^21.7b0"
mypy = "^0.910"
pylint = "^2.9.6"
isort = "^5.9.3"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
